+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.26.0] - 2026-08-14
+
+_Stable release based on [5.26.0-rc.1]._
+
+## [5.26.0-rc.1] - 2026-08-14
+
+### Added
+
+- Tag cloud Blade component, showing a list of terms as coloured chips or as a typographic word cloud fitted to the width it has.
+
 ## [5.25.1] - 2026-08-14
 
 ### Fixed
@@ -1832,6 +1842,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[5.26.0]: https://https://github.com/internetguru/laravel-common/compare/v5.25.1...v5.26.0
+[5.26.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.25.1
 [5.25.1]: https://https://github.com/internetguru/laravel-common/compare/v5.25.0...v5.25.1
 [5.25.0]: https://https://github.com/internetguru/laravel-common/compare/v5.24.0...v5.25.0
 [5.25.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.24.0